refactor: remove logic from component
